use a full-blown test package when testing
